AUTH

+++ 03_django/03_django_form/articles/views.py
@@ -1,5 +1,6 @@
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect, get_object_or_404, get_list_or_404
-from django.http import Http404
+from django.http import Http404, HttpResponse
 from django.views.decorators.http import require_POST
 from .models import Article, Comment
 from .forms import ArticleForm, CommentForm
@@ -9,12 +10,21 @@ from IPython import embed
 
 
 def index(request):
+    # embed()
+    # session 에 visits_num 키로 접근해 값을 가져온다. 
+    # 기본적으로 존재하지 않는 키이기 때문에 키가 없다면 (방문한 적이 없다면 ) 0 값을 가져오도록 한다.
+    visits_num = request.session.get('visits_num', 0)
+    # 그리고 가져온 값을 session 에 visits_num 에 매 번 1 씩 증가한 값으로 할당한다. 
+    request.session['visits_num'] = visits_num + 1
+    # django 에게 session data 커스텀 선언 
+    request.session.modified = True
+
     articles = Article.objects.all()
     #articles = get_list_or_404(Article) 글 개수가 0 이면 에러 나기 때문에 부적절
-    context = {'articles':articles}
+    context = {'articles':articles, 'visits_num': visits_num}
     return render(request, 'articles/index.html', context)
 
-
+@login_required
 def create(request):
 
     if request.method == 'POST':
@@ -43,12 +53,15 @@ def detail(request, article_pk):
     context = {'article':article, 'commentform':commentform, 'comments':comments}
     return render(request, 'articles/detail.html', context)
 
+
 @require_POST
 def delete(request, article_pk):
-    article = get_object_or_404(Article, pk=article_pk)
-    article.delete()
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=article_pk)
+        article.delete()
     return redirect('articles:index')
 
+@login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
     if request.method == 'POST':
@@ -66,23 +79,27 @@ def update(request, article_pk):
     context = {'form': form, 'article': article}
     return render(request, 'articles/form.html', context) 
 
-@require_POST
 
+
+@require_POST
 def comments_create(request, article_pk):
-    commentform = CommentForm(request.POST)
-    if commentform.is_valid():
-        # 객체 생성 하지만 DB에 레코드는 작성하지 않는다.
-        comment = commentform.save(commit=False)
-        comment.article_id = article_pk
-        comment.save()
+    if request.user.is_authenticated:
+        commentform = CommentForm(request.POST)
+        if commentform.is_valid():
+            # 객체 생성 하지만 DB에 레코드는 작성하지 않는다.
+            comment = commentform.save(commit=False)
+            comment.article_id = article_pk
+            comment.save()
     return redirect('articles:detail', article_pk)
 
 @require_POST
 def comments_delete(request, article_pk, comment_pk):
-    comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()
+    if request.user.is_authenticated:
+        comment = get_object_or_404(Comment, pk=comment_pk)
+        comment.delete()
 
         # return redirect('articles:detail' article.pk)
         # return redirect('articles:detail' article_pk)
 
-    return redirect('article:detail', article_pk)
+        return redirect('article:detail', article_pk)
+    return HttpResponse('You are unauthorized', status=401)
